refactor(ui): commonize source form modal (#299)
